Count pending bytes inside the io queue (#203)
shard_pool_fs reports the bytes its queued writes are still carrying: an
upper bound, since a write counts from the moment it is accepted, so the
figure can read high but never low. It used to be a subtraction of two
counters, one raised by the thread queueing a write and one by the
worker
that finished it. Two separate reads are not a consistent snapshot, and
the result was unsigned, so it could wrap to a number near 2^64. The
pool
also counted a write only after handing it to the worker, so the worker
could finish and lower the total before the raise landed.

The io queue keeps the count now. io_queue_post_bytes takes a byte count
and raises the total inside the same lock that puts the job in the
queue;
the worker lowers it inside the same lock that marks the job finished. A
reader can never see less work than the queue holds, so no ordering rule
is left to break. io_queue_post is that call with a count of zero,
leaving
jobs that carry no data unchanged.

shard_pool_fs and the throttled bench sink drop their own counters and
read
the queue's instead. The shard_pool and shard_sink headers state the
bound,
and pending_bytes is uint64_t rather than size_t throughout: both
interfaces, every implementation behind them, peak_pending_bytes, and
backpressure_bytes.

tests/test_io_queue.c covers the count directly. It blocks the worker,
checks the total after each post, then checks that it returns to zero
once
the queue drains and that a job posted without a byte count never
changes
it. The pool test checks that both write paths, copying and zero-copy,
report through. Tests now share one wait-for-flag helper instead of four
copies.

Closes #201.

---------

+++ bench/bench_util.c
@@ -525,7 +525,7 @@ run_bench(const struct bench_config* cfg)
                                   dtype_bpe(dtype),
                                   cfg->append_elements) == 0);
 
-  size_t pending_bytes = bench_zarr_pending_bytes(&zarr);
+  uint64_t pending_bytes = bench_zarr_pending_bytes(&zarr);
 
   struct platform_clock flush_clock = { 0 };
   platform_toc(&flush_clock);
@@ -1095,7 +1095,7 @@ run_bench_two_streams(const struct bench_config* cfg)
       print_metric_row(&m[k].backpressure);
       print_append_latency(&m[k]);
       char pbuf[32];
-      format_bytes(pbuf, sizeof(pbuf), (uint64_t)m[k].peak_pending_bytes);
+      format_bytes(pbuf, sizeof(pbuf), m[k].peak_pending_bytes);
       print_report("  peak pending:    %s", pbuf);
     }
   }
